# docs/swagger.yaml
definitions:
  controller.CommonError:
    properties:
      error:
        type: string
    type: object
  controller.CommonResponse:
    properties:
      message:
        type: string
    type: object
  controller.SignInRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  controller.SignUpRequest:
    properties:
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  controller.TokenPairResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  models.Employee:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact:
    email: test@test.com
    name: OnlineShop API Service
    url: http://test.com
  title: OnlineShop API
paths:
  /api/employees/:
    post:
      description: Функция добавления нового сотрудника
      parameters:
      - description: Информация о новом сотруднике
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: Добавление нового сотрудника
      tags:
      - Employees
  /api/employees/{id}:
    delete:
      description: Удаление сотрудника по ID
      parameters:
      - description: id сотрудника
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: Удалить сотрудника по ID
      tags:
      - Employees
    get:
      description: Получение конкретного сотрудника по его ID
      parameters:
      - description: id сотрудника
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: Получить сотрудника по ID
      tags:
      - Employees
    put:
      description: Обновление сотрудника с соответствующим ID
      parameters:
      - description: id сотрудника
        in: path
        name: id
        required: true
        type: integer
      - description: Информация о сотруднике
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: Обновить сотрудника по ID
      tags:
      - Employees
  /auth/refresh:
    get:
      description: Обновить пару токенов
      parameters:
      - description: вставьте refresh token
        in: header
        name: X-Refresh-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Обновить пару токенов
      tags:
      - Auth
  /auth/sign-in:
    post:
      description: Войти в аккаунт
      parameters:
      - description: логин и пароль
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Вход
      tags:
      - Auth
  /auth/sign-up:
    post:
      description: Создать новый аккаунт
      parameters:
      - description: информация о новом аккаунте
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/controller.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Регистрация
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"